tests/wpt/tests/html/canvas/element/layers/2d.layer.invalid-calls.beginLayer-save-endLayer.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.layer.invalid-calls.beginLayer-save-endLayer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.invalid-calls.beginLayer-save-endLayer</h1>
<p class="desc">Raises exception on beginLayer() + save() + endLayer().</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Raises exception on beginLayer() + save() + endLayer().");
_addTest(function(canvas, ctx) {

  assert_throws_dom("INVALID_STATE_ERR", function() {
    ctx.beginLayer();
    ctx.save();
    ctx.endLayer();
  });

});
</script>
